Fixed neverending loop

+++ src/main/java/es/us/isa/restest/generators/ALDrivenTestCaseGenerator.java
@@ -24,6 +24,8 @@ public class ALDrivenTestCaseGenerator extends AbstractTestCaseGenerator {
 	private String csvTmpTcPath; 									// resourcesFolderPath + "/" + CSV_NAME
 	private Integer numberOfCandidates; 							// number of candidates to perform ALT
 	private String queryStrategy; 									// Strategy to query best test cases among candidates
+	private Integer alMaxNumberOfTries;
+	private Integer iteration = 0;
 
 	private static Logger logger = LogManager.getLogger(ALDrivenTestCaseGenerator.class.getName());
 
@@ -64,49 +66,41 @@ public class ALDrivenTestCaseGenerator extends AbstractTestCaseGenerator {
 				queriedTestCases.add(test);
 			}
 
-			// Export test cases to temporary CSV
-			deleteFile(csvTmpTcPath); // Delete file first, so as to consider only test cases from this iteration
-			queriedTestCases.forEach(tc -> tc.exportToCSV(csvTmpTcPath));
+			if (iteration<alMaxNumberOfTries) {
+				// Max number of iterations not reached yet, execute the AL predictor
+				// Export test cases to temporary CSV
+				deleteFile(csvTmpTcPath); // Delete file first, so as to consider only test cases from this iteration
+				queriedTestCases.forEach(tc -> tc.exportToCSV(csvTmpTcPath));
 
-			// Feed test cases to predictor, which queries and labels the best ones
-			boolean commandOk = false;
-			try {
-				ProcessBuilder pb = new ProcessBuilder(alPredictorCommand, resourcesFolderPath, csvTmpTcPath, queryStrategy, ((Integer) numberOfTests).toString(), heedFaultyRatio.toString(), faultyRatio.toString());
-				pb.inheritIO(); // Print output of program to stdout
-				Process proc = pb.start();
-				proc.waitFor();
-				commandOk = true;
-			} catch (IOException e) {
-				logger.error("Error running AL selector");
-				logger.error("Exception: ", e);
-			} catch (InterruptedException e) {
-				logger.error("Error running AL selector");
-				logger.error("Exception: ", e);
-				Thread.currentThread().interrupt();
+				// Feed test cases to predictor, which queries and labels the best ones
+				boolean commandOk = false;
+				try {
+					ProcessBuilder pb = new ProcessBuilder(alPredictorCommand, resourcesFolderPath, csvTmpTcPath, queryStrategy, ((Integer) numberOfTests).toString(), heedFaultyRatio.toString(), faultyRatio.toString());
+					pb.inheritIO(); // Print output of program to stdout
+					Process proc = pb.start();
+					proc.waitFor();
+					commandOk = true;
+				} catch (IOException e) {
+					logger.error("Error running AL selector");
+					logger.error("Exception: ", e);
+				} catch (InterruptedException e) {
+					logger.error("Error running AL selector");
+					logger.error("Exception: ", e);
+					Thread.currentThread().interrupt();
+				}
+				if (commandOk) {
+					// Read back test cases from CSV and update objects
+					queriedTestCases = getTestCases(csvTmpTcPath);
+				}
+			} else {
+				heedFaultyRatio = false;
 			}
 
-			if (commandOk) {
-				// Read back test cases from CSV and update objects
-				queriedTestCases = getTestCases(csvTmpTcPath);
-
-				if (heedFaultyRatio) {
-					// Add test cases one by one until desired number is reached both for nominal and faulty
-					queriedTestCases.forEach(tc -> {
-						if ((Boolean.TRUE.equals(tc.getFaulty()) && hasNextFaulty()) ||
-								(Boolean.FALSE.equals(tc.getFaulty()) && hasNextNominal())) {
-							// Set authentication data (if any)
-							authenticateTestCase(tc);
-
-							// Add test case to the collection
-							testCases.add(tc);
-
-							// Update indexes
-							updateIndexes(tc);
-							Integer a = 1;
-						}
-					});
-				} else {
-					queriedTestCases.forEach(tc ->{
+			if (heedFaultyRatio) {
+				// Add test cases one by one until desired number is reached both for nominal and faulty
+				queriedTestCases.forEach(tc -> {
+					if ((Boolean.TRUE.equals(tc.getFaulty()) && hasNextFaulty()) ||
+							(Boolean.FALSE.equals(tc.getFaulty()) && hasNextNominal())) {
 						// Set authentication data (if any)
 						authenticateTestCase(tc);
 
@@ -115,13 +109,28 @@ public class ALDrivenTestCaseGenerator extends AbstractTestCaseGenerator {
 
 						// Update indexes
 						updateIndexes(tc);
-					});
-				}
+						Integer a = 1;
+					}
+				});
+			} else {
+				queriedTestCases.forEach(tc ->{
+					if (hasNext()) {
+						// Set authentication data (if any)
+						authenticateTestCase(tc);
+
+						// Add test case to the collection
+						testCases.add(tc);
+
+						// Update indexes
+						updateIndexes(tc);
+					}
+				});
 			}
+			iteration ++;
 		}
 		return testCases;
 	}
-	
+
 
 	// Generate the next test case
 	public TestCase generateNextTestCase(Operation testOperation) throws RESTestException {
@@ -159,5 +168,7 @@ public class ALDrivenTestCaseGenerator extends AbstractTestCaseGenerator {
 	public String getCsvTmpTcPath() { return csvTmpTcPath; }
 
 	public void setNumberOfCandidates(Integer numberOfCandidates) { this.numberOfCandidates = numberOfCandidates; }
+
+	public void setAlMaxNumberOfTries(Integer alMaxNumberOfTries) { this.alMaxNumberOfTries = alMaxNumberOfTries; }
 }
